Configs: imx6q-mf0300: disable HCD test mode driver
Typically this will be enabled for On-the-Go or embedded hosts that
need to undergo USB-IF compliance testing with the aid of special
testing hardware.  In the future, this may expand to include other
tests that require support from a HCD driver.
This option is of interest only to developers who need to validate
their USB hardware designs. It is not needed for normal use so we
just disable it.