DTS: imx6qdl-mf0300: use PLL5 clock as parent clock for LVDS
By default LDB_DI0_IPU and LDB_DI1_IPU clocks use MMDC_CH1 clock as
source. But it can't be adjusted enough to fulfill display requirements.
As result, LVDS bus can't transmit data fast enough.

PLL5 (PLL_VIDEO) clock is designed for usage in video subsystem and it
can be adjusted in wide range. So use it for LVDS.

Note: on Android 4.4 and Android 6 systems this clocks was used too.